Wrote getMotorBPM, getMotorIE() for motor_ctrl.cpp. Fixed return types in interface.cpp. Turns out the motor is attempting to turn, but the damn gearbox is slipping.

+++ BPAP/interface.h
@@ -7,9 +7,9 @@ int getROM();
 float getpressureThreshold();
 float getStepSize();
 
-int getBPM();
-int getVOL();
-int getIE();
+float getBPM();
+float getVOL();
+float getIE();
 int getASIST();
 int getALARM();
 float getSwitchToBag();
